smart-park-system/requirements/dev.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile '.\requirements\dev.in'
#
asgiref==3.9.1
    # via
    #   django
    #   django-cors-headers
asttokens==3.0.0
    # via stack-data
attrs==25.3.0
    # via
    #   jsonschema
    #   referencing
black==25.1.0
    # via -r requirements/dev.in
build==1.3.0
    # via pip-tools
certifi==2025.8.3
    # via requests
cfgv==3.4.0
    # via pre-commit
charset-normalizer==3.4.3
    # via requests
click==8.2.1
    # via
    #   black
    #   pip-tools
colorama==0.4.6
    # via
    #   build
    #   click
    #   ipython
    #   pytest
contourpy==1.3.3
    # via matplotlib
coverage[toml]==7.10.6
    # via
    #   -r requirements/dev.in
    #   pytest-cov
cycler==0.12.1
    # via matplotlib
decorator==5.2.1
    # via ipython
distlib==0.4.0
    # via virtualenv
django==5.2.6
    # via
    #   -r requirements/base.in
    #   django-cors-headers
    #   django-filter
    #   djangorestframework
    #   djangorestframework-simplejwt
    #   drf-spectacular
    #   model-bakery
django-cors-headers==4.8.0
    # via -r requirements/base.in
django-environ==0.11.2
    # via -r requirements/base.in
django-filter==24.3
    # via -r requirements/base.in
djangorestframework==3.15.2
    # via
    #   -r requirements/base.in
    #   djangorestframework-simplejwt
    #   drf-spectacular
djangorestframework-simplejwt==5.3.1
    # via -r requirements/base.in
drf-spectacular==0.27.2
    # via -r requirements/base.in
executing==2.2.1
    # via stack-data
filelock==3.19.1
    # via
    #   torch
    #   virtualenv
fonttools==4.59.2
    # via matplotlib
fsspec==2025.9.0
    # via torch
identify==2.6.14
    # via pre-commit
idna==3.10
    # via requests
inflection==0.5.1
    # via drf-spectacular
iniconfig==2.1.0
    # via pytest
ipython==9.5.0
    # via -r requirements/dev.in
ipython-pygments-lexers==1.1.1
    # via ipython
jedi==0.19.2
    # via ipython
jinja2==3.1.6
    # via torch
jsonschema==4.25.1
    # via drf-spectacular
jsonschema-specifications==2025.9.1
    # via jsonschema
kiwisolver==1.4.9
    # via matplotlib
markupsafe==3.0.2
    # via jinja2
matplotlib==3.10.6
    # via ultralytics
matplotlib-inline==0.1.7
    # via ipython
model-bakery==1.20.5
    # via -r requirements/dev.in
mpmath==1.3.0
    # via sympy
mypy==1.18.1
    # via -r requirements/dev.in
mypy-extensions==1.1.0
    # via
    #   black
    #   mypy
networkx==3.5
    # via torch
nodeenv==1.9.1
    # via pre-commit
numpy==2.2.6
    # via
    #   -r requirements/scripts.in
    #   contourpy
    #   matplotlib
    #   opencv-python
    #   scipy
    #   torchvision
    #   ultralytics
    #   ultralytics-thop
opencv-python==4.12.0.88
    # via
    #   -r requirements/scripts.in
    #   ultralytics
packaging==25.0
    # via
    #   black
    #   build
    #   matplotlib
    #   pytest
parso==0.8.5
    # via jedi
pathspec==0.12.1
    # via
    #   black
    #   mypy
pillow==11.3.0
    # via
    #   matplotlib
    #   torchvision
    #   ultralytics
pip-tools==7.5.0
    # via -r requirements/dev.in
platformdirs==4.4.0
    # via
    #   black
    #   virtualenv
pluggy==1.6.0
    # via
    #   pytest
    #   pytest-cov
polars==1.33.1
    # via ultralytics
pre-commit==4.3.0
    # via -r requirements/dev.in
prompt-toolkit==3.0.52
    # via ipython
psutil==7.0.0
    # via ultralytics
psycopg[binary]==3.2.10
    # via -r requirements/base.in
psycopg-binary==3.2.10
    # via psycopg
pure-eval==0.2.3
    # via stack-data
pydbml==1.2.0
    # via -r requirements/dev.in
pygments==2.19.2
    # via
    #   ipython
    #   ipython-pygments-lexers
    #   pytest
pyjwt==2.10.1
    # via djangorestframework-simplejwt
pyparsing==3.2.4
    # via
    #   matplotlib
    #   pydbml
pyproject-hooks==1.2.0
    # via
    #   build
    #   pip-tools
pytest==8.4.2
    # via
    #   -r requirements/dev.in
    #   pytest-cov
    #   pytest-django
pytest-cov==7.0.0
    # via -r requirements/dev.in
pytest-django==4.11.1
    # via -r requirements/dev.in
python-dateutil==2.9.0.post0
    # via matplotlib
pyyaml==6.0.2
    # via
    #   drf-spectacular
    #   pre-commit
    #   ultralytics
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.5
    # via
    #   -r requirements/scripts.in
    #   ultralytics
rpds-py==0.27.1
    # via
    #   jsonschema
    #   referencing
ruff==0.13.0
    # via -r requirements/dev.in
scipy==1.16.2
    # via ultralytics
six==1.17.0
    # via python-dateutil
sqlparse==0.5.3
    # via django
stack-data==0.6.3
    # via ipython
sympy==1.14.0
    # via torch
torch==2.8.0
    # via
    #   torchvision
    #   ultralytics
    #   ultralytics-thop
torchvision==0.23.0
    # via ultralytics
traitlets==5.14.3
    # via
    #   ipython
    #   matplotlib-inline
typing-extensions==4.15.0
    # via
    #   mypy
    #   torch
tzdata==2025.2
    # via
    #   django
    #   psycopg
ultralytics==8.3.200
    # via -r requirements/scripts.in
ultralytics-thop==2.0.17
    # via ultralytics
uritemplate==4.2.0
    # via drf-spectacular
urllib3==2.5.0
    # via requests
virtualenv==20.34.0
    # via pre-commit
wcwidth==0.2.13
    # via prompt-toolkit
wheel==0.45.1
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
